+++ mountpoint-s3-fs/src/superblock.rs
@@ -46,9 +46,7 @@ use crate::metablock::{InodeError, InodeKind, InodeNo, InodeStat, WriteMode};
 use crate::metablock::{InodeInformation, Lookup};
 use crate::metablock::{S3Location, TryAddDirEntry};
 use crate::metablock::{ValidKey, ValidName};
-use crate::prefix::Prefix;
-use crate::s3::S3Personality;
-use crate::s3::config::S3Path;
+use crate::s3::{S3Path, S3Personality};
 use crate::sync::atomic::{AtomicU64, Ordering};
 use crate::sync::{Arc, RwLock};
 
@@ -223,9 +221,9 @@ impl<'a> RenameLockGuard<'a> {
 
 impl<OC: ObjectClient + Send + Sync> Superblock<OC> {
     /// Create a new Superblock that targets the given bucket/prefix
-    pub fn new(client: OC, bucket: &str, prefix: &Prefix, config: SuperblockConfig) -> Self {
+    pub fn new(client: OC, s3_path: S3Path, config: SuperblockConfig) -> Self {
         let mount_time = OffsetDateTime::now_utc();
-        let root = Inode::new_root(prefix, mount_time);
+        let root = Inode::new_root(&s3_path.prefix, mount_time);
 
         let mut inodes = InodeMap::default();
         inodes.insert(root.ino(), root, 1);
@@ -236,10 +234,7 @@ impl<OC: ObjectClient + Send + Sync> Superblock<OC> {
         );
 
         let inner = SuperblockInner {
-            s3_path: Arc::new(S3Path {
-                bucket_name: bucket.to_owned(),
-                prefix: prefix.clone(),
-            }),
+            s3_path: Arc::new(s3_path),
             inodes: RwLock::new(inodes),
             reader_counts: Default::default(),
             negative_cache,
@@ -423,12 +418,7 @@ impl<OC: ObjectClient + Send + Sync + Clone> Metablock for Superblock<OC> {
         let rename_object_result = self
             .inner
             .client
-            .rename_object(
-                &self.inner.s3_path.bucket_name,
-                src_key.as_ref(),
-                &dest_key,
-                &rename_params,
-            )
+            .rename_object(&self.inner.s3_path.bucket, src_key.as_ref(), &dest_key, &rename_params)
             .await;
 
         match rename_object_result {
@@ -460,7 +450,7 @@ impl<OC: ObjectClient + Send + Sync + Clone> Metablock for Superblock<OC> {
                     _ => Err(InodeError::client_error(
                         error,
                         "RenameObject failed",
-                        &self.inner.s3_path.bucket_name,
+                        &self.inner.s3_path.bucket,
                         src_key.as_ref(),
                     )),
                 };
@@ -637,7 +627,7 @@ impl<OC: ObjectClient + Send + Sync + Clone> Metablock for Superblock<OC> {
                 return Err(InodeError::UnlinkNotPermittedWhileWriting(inode.err()));
             }
             WriteStatus::Remote => {
-                let bucket = &self.inner.s3_path.bucket_name;
+                let bucket = &self.inner.s3_path.bucket;
                 let s3_key = self.inner.full_key_for_inode(&inode);
                 debug!(parent=?parent_ino, ?name, "unlink on remote file will delete key {}", s3_key);
                 let delete_obj_result = self.inner.client.delete_object(bucket, &s3_key).await;
@@ -1288,11 +1278,11 @@ impl<OC: ObjectClient + Send + Sync> SuperblockInner<OC> {
         let head_object_params = HeadObjectParams::new();
         let mut file_lookup = self
             .client
-            .head_object(&self.s3_path.bucket_name, object_key, &head_object_params)
+            .head_object(&self.s3_path.bucket, object_key, &head_object_params)
             .fuse();
         let mut dir_lookup = self
             .client
-            .list_objects(&self.s3_path.bucket_name, None, "/", 1, directory_prefix)
+            .list_objects(&self.s3_path.bucket, None, "/", 1, directory_prefix)
             .fuse();
 
         let mut file_state = None;
@@ -1307,12 +1297,12 @@ impl<OC: ObjectClient + Send + Sync> SuperblockInner<OC> {
                         }
                         // If the object is not found, might be a directory, so keep going
                         Err(ObjectClientError::ServiceError(HeadObjectError::NotFound)) => {},
-                        Err(e) => return Err(InodeError::client_error(e, "HeadObject failed", &self.s3_path.bucket_name, object_key)),
+                        Err(e) => return Err(InodeError::client_error(e, "HeadObject failed", &self.s3_path.bucket, object_key)),
                     }
                 }
 
                 result = dir_lookup => {
-                    let result = result.map_err(|e| InodeError::client_error(e, "ListObjectsV2 failed", &self.s3_path.bucket_name, object_key))?;
+                    let result = result.map_err(|e| InodeError::client_error(e, "ListObjectsV2 failed", &self.s3_path.bucket, object_key))?;
 
                     let found_directory = if result
                         .common_prefixes
@@ -1806,6 +1796,7 @@ mod tests {
     use time::{Duration, OffsetDateTime};
 
     use crate::fs::{FUSE_ROOT_INODE, TimeToLive, ToErrno};
+    use crate::s3::{Bucket, Prefix};
 
     use super::*;
 
@@ -1824,8 +1815,13 @@ mod tests {
     #[test_case("test_prefix/"; "prefixed")]
     #[tokio::test]
     async fn test_lookup(prefix: &str) {
-        let bucket = "test_bucket";
-        let client = Arc::new(MockClient::config().bucket(bucket).part_size(1024 * 1024).build());
+        let bucket = Bucket::new("test_bucket").unwrap();
+        let client = Arc::new(
+            MockClient::config()
+                .bucket(bucket.to_string())
+                .part_size(1024 * 1024)
+                .build(),
+        );
 
         let keys = &[
             format!("{prefix}dir0/file0.txt"),
@@ -1852,7 +1848,11 @@ mod tests {
 
         let prefix = Prefix::new(prefix).expect("valid prefix");
         let ts = OffsetDateTime::now_utc();
-        let superblock = Superblock::new(client.clone(), bucket, &prefix, Default::default());
+        let superblock = Superblock::new(
+            client.clone(),
+            S3Path::new(bucket.clone(), prefix.clone()),
+            Default::default(),
+        );
 
         // Try it twice to test the inode reuse path too
         for _ in 0..2 {
@@ -1930,7 +1930,7 @@ mod tests {
                     // Grab last modified time according to mock S3
                     let full_key = file.s3_location().expect("should have location").full_key();
                     let modified_time = client
-                        .head_object(bucket, full_key.as_ref(), &HeadObjectParams::new())
+                        .head_object(&bucket, full_key.as_ref(), &HeadObjectParams::new())
                         .await
                         .expect("object should exist")
                         .last_modified;
@@ -1945,9 +1945,14 @@ mod tests {
     #[test_case(false; "not cached")]
     #[tokio::test]
     async fn test_lookup_with_caching(cached: bool) {
-        let bucket = "test_bucket";
+        let bucket = Bucket::new("test_bucket").unwrap();
         let prefix = "prefix/";
-        let client = Arc::new(MockClient::config().bucket(bucket).part_size(1024 * 1024).build());
+        let client = Arc::new(
+            MockClient::config()
+                .bucket(bucket.to_string())
+                .part_size(1024 * 1024)
+                .build(),
+        );
 
         let keys = &[
             format!("{prefix}file0.txt"),
@@ -1972,8 +1977,7 @@ mod tests {
         };
         let superblock = Superblock::new(
             client.clone(),
-            bucket,
-            &prefix,
+            S3Path::new(bucket, prefix.clone()),
             SuperblockConfig {
                 cache_config: CacheConfig::new(TimeToLive::Duration(ttl)),
                 s3_personality: S3Personality::Standard,
@@ -2006,9 +2010,9 @@ mod tests {
     #[test_case(false; "not cached")]
     #[tokio::test]
     async fn test_negative_lookup_with_caching(cached: bool) {
-        let bucket = "test_bucket";
+        let bucket = Bucket::new("test_bucket").unwrap();
         let prefix = "prefix/";
-        let client = Arc::new(MockClient::config().bucket(bucket).part_size(32).build());
+        let client = Arc::new(MockClient::config().bucket(bucket.to_string()).part_size(32).build());
 
         let prefix = Prefix::new(prefix).expect("valid prefix");
         let ttl = if cached {
@@ -2018,8 +2022,7 @@ mod tests {
         };
         let superblock = Superblock::new(
             client.clone(),
-            bucket,
-            &prefix,
+            S3Path::new(bucket, prefix.clone()),
             SuperblockConfig {
                 cache_config: CacheConfig::new(TimeToLive::Duration(ttl)),
                 s3_personality: S3Personality::Standard,
@@ -2063,9 +2066,10 @@ mod tests {
     #[test_case("test_prefix/"; "prefixed")]
     #[tokio::test]
     async fn test_readdir(prefix: &str) {
+        let bucket = Bucket::new("test_bucket").unwrap();
         let client = Arc::new(
             MockClient::config()
-                .bucket("test_bucket")
+                .bucket(bucket.to_string())
                 .part_size(1024 * 1024)
                 .build(),
         );
@@ -2093,7 +2097,7 @@ mod tests {
 
         let prefix = Prefix::new(prefix).expect("valid prefix");
         let ts = OffsetDateTime::now_utc();
-        let superblock = Superblock::new(client.clone(), "test_bucket", &prefix, Default::default());
+        let superblock = Superblock::new(client.clone(), S3Path::new(bucket, prefix.clone()), Default::default());
 
         // Try it all twice to test inode reuse
         for _ in 0..2 {
@@ -2218,10 +2222,11 @@ mod tests {
     #[test_case("test_prefix/"; "prefixed")]
     #[tokio::test]
     async fn test_readdir_no_remote_keys(prefix: &str) {
-        let client = Arc::new(MockClient::config().bucket("test_bucket").part_size(32).build());
+        let bucket = Bucket::new("test_bucket").unwrap();
+        let client = Arc::new(MockClient::config().bucket(bucket.to_string()).part_size(32).build());
 
         let prefix = Prefix::new(prefix).expect("valid prefix");
-        let superblock = Superblock::new(client.clone(), "test_bucket", &prefix, Default::default());
+        let superblock = Superblock::new(client.clone(), S3Path::new(bucket, prefix.clone()), Default::default());
 
         let mut expected_list: Vec<OsString> = Vec::new();
 
@@ -2250,15 +2255,16 @@ mod tests {
     #[test_case("test_prefix/"; "prefixed")]
     #[tokio::test]
     async fn test_readdir_local_keys_after_remote_keys(prefix: &str) {
+        let bucket = Bucket::new("test_bucket").unwrap();
         let client = Arc::new(
             MockClient::config()
-                .bucket("test_bucket")
+                .bucket(bucket.to_string())
                 .part_size(1024 * 1024)
                 .build(),
         );
 
         let prefix = Prefix::new(prefix).expect("valid prefix");
-        let superblock = Superblock::new(client.clone(), "test_bucket", &prefix, Default::default());
+        let superblock = Superblock::new(client.clone(), S3Path::new(bucket, prefix.clone()), Default::default());
 
         let mut expected_list: Vec<OsString> = Vec::new();
 
@@ -2298,14 +2304,15 @@ mod tests {
     #[test_case("test_prefix/"; "prefixed")]
     #[tokio::test]
     async fn test_create_local_dir(prefix: &str) {
+        let bucket = Bucket::new("test_bucket").unwrap();
         let client = Arc::new(
             MockClient::config()
-                .bucket("test_bucket")
+                .bucket(bucket.to_string())
                 .part_size(1024 * 1024)
                 .build(),
         );
         let prefix = Prefix::new(prefix).expect("valid prefix");
-        let superblock = Superblock::new(client.clone(), "test_bucket", &prefix, Default::default());
+        let superblock = Superblock::new(client.clone(), S3Path::new(bucket, prefix.clone()), Default::default());
 
         // Create local directory
         let dirname = "local_dir";
@@ -2335,14 +2342,15 @@ mod tests {
     #[test_case("test_prefix/"; "prefixed")]
     #[tokio::test]
     async fn test_readdir_lookup_after_rmdir(prefix: &str) {
+        let bucket = Bucket::new("test_bucket").unwrap();
         let client = Arc::new(
             MockClient::config()
-                .bucket("test_bucket")
+                .bucket(bucket.to_string())
                 .part_size(1024 * 1024)
                 .build(),
         );
         let prefix = Prefix::new(prefix).expect("valid prefix");
-        let superblock = Superblock::new(client.clone(), "test_bucket", &prefix, Default::default());
+        let superblock = Superblock::new(client.clone(), S3Path::new(bucket, prefix.clone()), Default::default());
 
         // Create local directory
         let dirname = "local_dir";
@@ -2391,22 +2399,12 @@ mod tests {
     #[test_case("test_prefix/", false; "prefixed unordered")]
     #[tokio::test]
     async fn test_readdir_unordered(prefix: &str, ordered: bool) {
-        let client = if ordered {
-            Arc::new(
-                MockClient::config()
-                    .bucket("test_bucket")
-                    .part_size(1024 * 1024)
-                    .build(),
-            )
-        } else {
-            Arc::new(
-                MockClient::config()
-                    .bucket("test_bucket")
-                    .part_size(1024 * 1024)
-                    .unordered_list_seed(Some(123456))
-                    .build(),
-            )
-        };
+        let bucket = Bucket::new("test_bucket").unwrap();
+        let mut config = MockClient::config().bucket(bucket.to_string()).part_size(1024 * 1024);
+        if !ordered {
+            config = config.unordered_list_seed(Some(123456));
+        }
+        let client = Arc::new(config.build());
 
         let prefix = Prefix::new(prefix).expect("valid prefix");
         let s3_personality = if ordered {
@@ -2416,8 +2414,7 @@ mod tests {
         };
         let superblock = Superblock::new(
             client.clone(),
-            "test_bucket",
-            &prefix,
+            S3Path::new(bucket, prefix.clone()),
             SuperblockConfig {
                 s3_personality,
                 ..Default::default()
@@ -2519,14 +2516,15 @@ mod tests {
     #[test_case("test_prefix/"; "prefixed")]
     #[tokio::test]
     async fn test_rmdir_delete_status(prefix: &str) {
+        let bucket = Bucket::new("test_bucket").unwrap();
         let client = Arc::new(
             MockClient::config()
-                .bucket("test_bucket")
+                .bucket(bucket.to_string())
                 .part_size(1024 * 1024)
                 .build(),
         );
         let prefix = Prefix::new(prefix).expect("valid prefix");
-        let superblock = Superblock::new(client.clone(), "test_bucket", &prefix, Default::default());
+        let superblock = Superblock::new(client.clone(), S3Path::new(bucket, prefix.clone()), Default::default());
 
         // Create local directory
         let dirname = "local_dir";
@@ -2571,14 +2569,15 @@ mod tests {
     #[test_case("test_prefix/"; "prefixed")]
     #[tokio::test]
     async fn test_parent_readdir_after_rmdir(prefix: &str) {
+        let bucket = Bucket::new("test_bucket").unwrap();
         let client = Arc::new(
             MockClient::config()
-                .bucket("test_bucket")
+                .bucket(bucket.to_string())
                 .part_size(1024 * 1024)
                 .build(),
         );
         let prefix = Prefix::new(prefix).expect("valid prefix");
-        let superblock = Superblock::new(client.clone(), "test_bucket", &prefix, Default::default());
+        let superblock = Superblock::new(client.clone(), S3Path::new(bucket, prefix.clone()), Default::default());
 
         // Create local directory
         let dirname = "local_dir";
@@ -2608,14 +2607,15 @@ mod tests {
     #[test_case("test_prefix/"; "prefixed")]
     #[tokio::test]
     async fn test_lookup_after_unlink(prefix: &str) {
+        let bucket = Bucket::new("test_bucket").unwrap();
         let client = Arc::new(
             MockClient::config()
-                .bucket("test_bucket".to_string())
+                .bucket(bucket.to_string())
                 .part_size(1024 * 1024)
                 .build(),
         );
         let prefix = Prefix::new(prefix).expect("valid prefix");
-        let superblock = Superblock::new(client.clone(), "test_bucket", &prefix, Default::default());
+        let superblock = Superblock::new(client.clone(), S3Path::new(bucket, prefix.clone()), Default::default());
 
         let file_name = "file.txt";
         let file_key = format!("{prefix}{file_name}");
@@ -2642,13 +2642,18 @@ mod tests {
 
     #[tokio::test]
     async fn test_finish_writing_convert_parent_local_dirs_to_remote() {
+        let bucket = Bucket::new("test_bucket").unwrap();
         let client = Arc::new(
             MockClient::config()
-                .bucket("test_bucket".to_string())
+                .bucket(bucket.to_string())
                 .part_size(1024 * 1024)
                 .build(),
         );
-        let superblock = Superblock::new(client.clone(), "test_bucket", &Default::default(), Default::default());
+        let superblock = Superblock::new(
+            client.clone(),
+            S3Path::new(bucket, Default::default()),
+            Default::default(),
+        );
 
         let nested_dirs = (0..5).map(|i| format!("level{i}")).collect::<Vec<_>>();
         let leaf_dir_ino = {
@@ -2695,15 +2700,20 @@ mod tests {
 
     #[tokio::test]
     async fn test_inode_reuse() {
+        let bucket = Bucket::new("test_bucket").unwrap();
         let client = Arc::new(
             MockClient::config()
-                .bucket("test_bucket".to_string())
+                .bucket(bucket.to_string())
                 .part_size(1024 * 1024)
                 .build(),
         );
         client.add_object("dir1/file1.txt", MockObject::constant(0xaa, 30, ETag::for_tests()));
 
-        let superblock = Superblock::new(client.clone(), "test_bucket", &Default::default(), Default::default());
+        let superblock = Superblock::new(
+            client.clone(),
+            S3Path::new(bucket, Default::default()),
+            Default::default(),
+        );
 
         for _ in 0..2 {
             let dir1_1 = superblock.lookup(FUSE_ROOT_INODE, "dir1".as_ref()).await.unwrap();
@@ -2720,9 +2730,10 @@ mod tests {
     #[test_case("subdir/"; "with subdirectory")]
     #[tokio::test]
     async fn test_lookup_directory_overlap(subdir: &str) {
+        let bucket = Bucket::new("test_bucket").unwrap();
         let client = Arc::new(
             MockClient::config()
-                .bucket("test_bucket".to_string())
+                .bucket(bucket.to_string())
                 .part_size(1024 * 1024)
                 .build(),
         );
@@ -2740,7 +2751,11 @@ mod tests {
             MockObject::constant(0xaa, 30, ETag::for_tests()),
         );
 
-        let superblock = Superblock::new(client.clone(), "test_bucket", &Default::default(), Default::default());
+        let superblock = Superblock::new(
+            client.clone(),
+            S3Path::new(bucket, Default::default()),
+            Default::default(),
+        );
 
         let entries = collect_dir_entries(&superblock, FUSE_ROOT_INODE, false, 2).await;
         assert_eq!(entries, &["dir", "dir-1"]);
@@ -2754,9 +2769,10 @@ mod tests {
 
     #[tokio::test]
     async fn test_invalid_names() {
+        let bucket = Bucket::new("test_bucket").unwrap();
         let client = Arc::new(
             MockClient::config()
-                .bucket("test_bucket")
+                .bucket(bucket.to_string())
                 .part_size(1024 * 1024)
                 .build(),
         );
@@ -2784,7 +2800,11 @@ mod tests {
             MockObject::constant(0xaa, 30, ETag::from_str("test_etag_5").unwrap()),
         );
 
-        let superblock = Superblock::new(client.clone(), "test_bucket", &Default::default(), Default::default());
+        let superblock = Superblock::new(
+            client.clone(),
+            S3Path::new(bucket, Default::default()),
+            Default::default(),
+        );
         let entries = collect_dir_entries_with_info(&superblock, FUSE_ROOT_INODE, true, 2).await;
         assert_eq!(entries.iter().map(|(_info, name)| name).collect::<Vec<_>>(), &["dir1"]);
 
@@ -2803,14 +2823,15 @@ mod tests {
     #[test_case("test_prefix/"; "prefixed")]
     #[tokio::test]
     async fn test_setattr(prefix: &str) {
+        let bucket = Bucket::new("test_bucket").unwrap();
         let client = Arc::new(
             MockClient::config()
-                .bucket("test_bucket")
+                .bucket(bucket.to_string())
                 .part_size(1024 * 1024)
                 .build(),
         );
         let prefix = Prefix::new(prefix).expect("valid prefix");
-        let superblock = Superblock::new(client.clone(), "test_bucket", &prefix, Default::default());
+        let superblock = Superblock::new(client.clone(), S3Path::new(bucket, prefix.clone()), Default::default());
 
         // Create a new file
         let filename = "newfile.txt";