+++ mountpoint-s3/src/fstab.rs
@@ -215,7 +215,7 @@ mod tests {
             match expected_prefix {
                 Ok(prefix) => {
                     let s3_path = cli_args.s3_path().unwrap();
-                    assert_eq!(s3_path.bucket_name, "demo_s3_bucket");
+                    assert_eq!(s3_path.bucket.as_str(), "demo_s3_bucket");
                     assert_eq!(s3_path.prefix.as_str(), prefix.as_str());
                 }
                 Err(expected_err) => {
@@ -246,7 +246,7 @@ mod tests {
 
         assert!(!cli_args.foreground);
         assert_eq!(cli_args.mount_point.to_str(), Some("/mnt/test"));
-        assert_eq!(s3_path.bucket_name, "demo_s3_bucket");
+        assert_eq!(s3_path.bucket.as_str(), "demo_s3_bucket");
         assert_eq!(s3_path.prefix.as_str(), "foo/bar,baz/");
         assert_eq!(cli_args.uid, Some(2));
         assert_eq!(cli_args.gid, None);
@@ -323,7 +323,7 @@ mod tests {
     impl From<CliArgs> for FstabCompatibleCliArgs {
         fn from(cli: CliArgs) -> Self {
             FstabCompatibleCliArgs {
-                bucket_name: cli.s3_path().unwrap().bucket_name,
+                bucket_name: cli.s3_path().unwrap().bucket.to_string(),
                 mount_point: cli.mount_point.to_string_lossy().into_owned(),
                 uid: cli.uid.unwrap_or_default(),
                 allow_delete: cli.allow_delete,