+++ mountpoint-s3-fs/CHANGELOG.md
@@ -1,4 +1,7 @@
-## Unreleased
+## Unreleased (v0.7.0)
+
+* Replace `S3Uri` with `S3Path` and consolidate related types like `Bucket` and `Prefix` into the `s3` module.
+  ([#1535](https://github.com/awslabs/mountpoint-s3/pull/1535))
 
 ## v0.6.0 (July 23, 2025)